requirements.txt
arrow==1.3.0
asgiref==3.9.1
attrs==25.3.0
binaryornot==0.4.4
certifi==2025.7.14
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cookiecutter==2.6.0
Django==5.2.4
django-ckeditor==6.7.3
django-cors-headers==4.7.0
django-js-asset==3.1.2
django-tailwind==4.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pillow==11.3.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20250708
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0

